Restructured backend to separate test files from source files. Added __init__.py files where necessary so PyTest can read packages appropriately. Finalized initial implementation of BitdotioDB.py. Wrote some initial unit tests for database calls.
