--- backend/Constants.py
@@ -1,2 +1,0 @@
-DB_NAME = 'equidistant'
-DB_TEST = 'test'
