+++ backend/src/Constants.py
@@ -1,0 +1,2 @@
+DB_NAME = 'nelsonys/equidistant'
+DB_TEST = 'test'
