+++ backend/tests/test_basic.py
@@ -1,2 +1,6 @@
 def test_something():
-    assert (True and True) == (True or True)
+    """
+    Simply verifies that PyTest is running as it should. Something is
+    seriously wrong with the repo setup if this test is failing.
+    """
+    assert True
