+++ backend/tests/test_db.py
@@ -1,23 +1,14 @@
-import unittest
-from backend.database.BitdotioDB import BitdotioDB
-from backend import Constants
+from backend.src import Constants
+from backend.src.database.BitdotioDB import BitdotioDB
 
 
-class TestDatabase(unittest.TestCase):
+def test_access_table():
+    db = BitdotioDB(token='v2_3ywUP_6X5W3jrBqMZQx7Vkyp9DAdN')
+    tuples = db.db_query(f'DELETE FROM {Constants.DB_TEST}')  # delete all data from the test table
+    assert len(tuples) == 0
+    tuples = db.db_query(f'INSERT INTO {Constants.DB_TEST} VALUES (\'some_string\', 42)')
+    assert len(tuples) == 0
+    tuples = db.db_query(f'SELECT * FROM {Constants.DB_TEST}')
+    assert len(tuples) == 1
+    assert tuples[0] == ('some_string', 42)
 
-    def test_test_table(self):
-        db = BitdotioDB(token='v2_3yqLc_FLiJXrBwLYR65zHT5shxqNs')
-        db.db_query(f'DELETE FROM {Constants.DB_TEST}')  # delete all data from the test table
-        tuples = db.db_query(f'INSERT INTO Test VALUES (\'some_string\', 42)')
-        self.assertEqual(len(tuples), 0)
-        tuples = db.db_query(f'SELECT * FROM Test')
-        self.assertEqual(len(tuples), 1)
-        self.assertEqual(tuples[0], 'some_string')
-        self.assertEqual(tuples[1], 42)
-
-    def test_raises_error(self):
-        pass
-
-
-if __name__ == '__main__':
-    unittest.main()
